Move add_safe_directory_on_remote tests to hosts/common_test.py
Co-locate tests with the function they test, which was moved to
hosts/common.py in the previous commit.
