+++ libs/mng/imbue/mng/utils/git_utils_test.py
@@ -2,12 +2,8 @@
 
 import subprocess
 from pathlib import Path
-from typing import cast
 
 from imbue.concurrency_group.concurrency_group import ConcurrencyGroup
-from imbue.mng.api.test_fixtures import FakeHost
-from imbue.mng.hosts.common import add_safe_directory_on_remote
-from imbue.mng.interfaces.host import OnlineHostInterface
 from imbue.mng.utils.git_utils import _parse_project_name_from_url
 from imbue.mng.utils.git_utils import derive_project_name_from_path
 from imbue.mng.utils.git_utils import find_git_common_dir
@@ -276,45 +272,3 @@ def test_get_git_remote_url_returns_none_for_non_git_dir(tmp_path: Path, cg: Con
     plain_dir = tmp_path / "plain"
     plain_dir.mkdir()
     assert get_git_remote_url(plain_dir, "origin", cg) is None
-
-
-# =============================================================================
-# add_safe_directory_on_remote tests
-# =============================================================================
-
-
-def _get_safe_directories(tmp_path: Path) -> list[str]:
-    """Read safe.directory entries from the gitconfig in the test HOME."""
-    gitconfig = tmp_path / ".gitconfig"
-    if not gitconfig.exists():
-        return []
-    result = subprocess.run(
-        ["git", "config", "--global", "--get-all", "safe.directory"],
-        capture_output=True,
-        text=True,
-    )
-    if result.returncode != 0:
-        return []
-    return result.stdout.strip().splitlines()
-
-
-def test_add_safe_directory_on_remote_adds_entry_for_non_local_host(tmp_path: Path) -> None:
-    """Test that add_safe_directory_on_remote writes to gitconfig for non-local hosts."""
-    host = cast(OnlineHostInterface, FakeHost(is_local=False))
-    target_path = Path("/some/agent/workdir")
-
-    add_safe_directory_on_remote(host, target_path)
-
-    safe_dirs = _get_safe_directories(tmp_path)
-    assert str(target_path) in safe_dirs
-
-
-def test_add_safe_directory_on_remote_is_noop_for_local_host(tmp_path: Path) -> None:
-    """Test that add_safe_directory_on_remote does nothing for local hosts."""
-    host = cast(OnlineHostInterface, FakeHost(is_local=True))
-    target_path = Path("/some/agent/workdir")
-
-    add_safe_directory_on_remote(host, target_path)
-
-    safe_dirs = _get_safe_directories(tmp_path)
-    assert str(target_path) not in safe_dirs
